+++ ckanext/saml/const.py
@@ -1,4 +1,4 @@
-MAP = {
+DEFAULT_MAPPING = {
     "user_id": "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier",
     "email": "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress",
     "name": "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/name",
@@ -7,3 +7,6 @@ MAP = {
     "upn": "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/upn",
     "groups": "http://schemas.xmlsoap.org/claims/Group",
 }
+
+MAPPING_PREFIX: str = "ckanext.saml.mapping."
+SETTINGS_PREFIX: str = "ckanext.saml.settings."