+++ ckanext/saml/migration/saml/versions/e8e7ebedf90d_add_attributes_column_to_saml2_user_.py
@@ -5,8 +5,8 @@ Revises: 25dc326c059e
 Create Date: 2022-08-01 15:55:55.408354
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 from sqlalchemy.dialects.postgresql import JSONB
 
 # revision identifiers, used by Alembic.