+++ ckanext/saml/tests/conftest.py
@@ -1,6 +1,11 @@
+import os
+
 import pytest
-from ckan.tests import factories
+from onelogin.saml2.idp_metadata_parser import \
+    OneLogin_Saml2_IdPMetadataParser as Parser
 from pytest_factoryboy import register
+
+from ckan.tests import factories
 
 
 @pytest.fixture
@@ -12,3 +17,11 @@ def clean_db(reset_db, migrate_db_for):
 @register
 class UserFactory(factories.User):
     pass
+
+
+@pytest.fixture
+def idp_metadata():
+    here = os.path.dirname(__file__)
+
+    with open(os.path.join(here, "data", "metadata_example.xml")) as f:
+        return Parser.parse(f.read())["idp"]