+++ ckanext/saml/tests/test_plugin.py
@@ -1,6 +1,0 @@
-"""Tests for plugin.py."""
-import ckanext.saml.plugin as plugin
-
-
-def test_plugin():
-    pass