+++ ckanext/saml/utils.py
@@ -9,24 +9,24 @@ from onelogin.saml2.auth import OneLogin_Saml2_Auth, OneLogin_Saml2_Utils
 import ckan.plugins as plugins
 import ckan.plugins.toolkit as tk
 
-from .interfaces import ICKANSAML
-from . import config
+from ckanext.saml import config
+from ckanext.saml.interfaces import ICKANSAML
 
 
 def prepare_from_flask_request() -> dict[str, Any]:
     url_data = urlparse(tk.request.url)
-
     req_path = tk.request.path
-    if tk.asbool(tk.config.get("ckan.saml_use_root_path", False)):
-        # FIX FOR ROOT_PATH REMOVED IN request.path
-        root_path = tk.config.get("ckan.root_path", None)
-        if root_path:
-            root_path = re.sub("/{{LANG}}", "", root_path)
-            req_path = root_path + req_path
-
     host = tk.request.host
     static_host = config.static_host()
     forwarded_host = tk.request.environ.get("HTTP_X_FORWARDED_HOST")
+
+    if config.use_root_path():
+        # FIX FOR ROOT_PATH REMOVED IN request.path
+        root_path: str = tk.config["ckan.root_path"]
+
+        if root_path:
+            root_path = re.sub("/{{LANG}}", "", root_path)
+            req_path = root_path + req_path
 
     if config.use_forwarded_host() and forwarded_host:
         host = forwarded_host
@@ -34,7 +34,7 @@ def prepare_from_flask_request() -> dict[str, Any]:
         host = static_host
 
     return {
-        "https": config.https(),
+        "https": config.use_https(),
         "http_host": host,
         "server_port": url_data.port,
         "script_name": req_path,
@@ -51,11 +51,7 @@ def make_auth(req: dict[str, Any]) -> OneLogin_Saml2_Auth:
     else:
         Auth = OneLogin_Saml2_Auth
 
-    if config.use_dynamic_config():
-        return Auth(req, old_settings=tk.h.saml_settings())
-
-    custom_folder = tk.h.saml_folder_path()
-    return Auth(req, custom_base_path=custom_folder)
+    return Auth(req, tk.h.saml_get_settings())
 
 
 def decode_saml_response(response: str) -> bytes: